Move OnionShareGui into its own file, so __init__ is just GUI logic
